requirements.txt
Flask==2.3.3
oracledb==3.4.0
python-dotenv==1.0.0
Werkzeug==2.3.7
